Remove undesired styles, content, and images
This is the first commit in a new branch. This projject was originally
built in a group for a scholastic assignment, but after the course ended
I decided to branch this repo and continue working on it on my own. I
cleaned up the content and styles that had been included by a previous
group member who was just beginning to explore the world of bootstrap
and css. This is now a "blank" version of the project we completed with
all the requirements still in tact.